+++ test/controllers/folio/console/links_controller_test.rb
@@ -2,26 +2,26 @@
 
 require 'test_helper'
 
-class Folio::Console::LinksControllerTest < Folio::Console::BaseControllerTest
+class Folio::Console::Api::LinksControllerTest < Folio::Console::BaseControllerTest
   include Folio::Engine.routes.url_helpers
 
   test 'index' do
     I18n.with_locale(:cs) do
-      get console_links_path
-      assert_equal([], JSON.parse(response.body))
+      get console_api_links_path
+      assert_equal([], JSON.parse(response.body)['data'])
 
       page = create(:folio_page, title: 'Foo', slug: 'foo')
-      get console_links_path
+      get console_api_links_path
       assert_equal([{ 'name' => 'Stránka - Foo', 'url' => '/foo' }],
-                   JSON.parse(response.body))
+                   JSON.parse(response.body)['data'])
 
       create(:folio_page, title: 'Bar', slug: 'bar')
-      get console_links_path
+      get console_api_links_path
       assert_equal([{ 'name' => 'Stránka - Bar', 'url' => '/bar' },
                     { 'name' => 'Stránka - Foo', 'url' => '/foo' }],
-                   JSON.parse(response.body))
+                   JSON.parse(response.body)['data'])
 
-      Folio::Console::LinksController.class_eval do
+      Folio::Console::Api::LinksController.class_eval do
         private
 
           def additional_links
@@ -39,12 +39,12 @@ class Folio::Console::LinksControllerTest < Folio::Console::BaseControllerTest
 
       create(:folio_menu_item, title: 'Test', target: page)
 
-      get console_links_path
+      get console_api_links_path
       assert_equal([{ 'name' => 'Homepage', 'url' => '/' },
                     { 'name' => 'Odkaz - Test', 'url' => 'url' },
                     { 'name' => 'Stránka - Bar', 'url' => '/bar' },
                     { 'name' => 'Stránka - Foo', 'url' => '/foo' }],
-                   JSON.parse(response.body))
+                   JSON.parse(response.body)['data'])
     end
   end
 end